feat: migrate from Turso/SQLite to Neon Postgres
Replaces the dual SQLite/Turso adapter setup with a single Neon Postgres
connection. Migrations are now applied via standard `prisma migrate deploy`
at build time, removing the custom Turso migration script entirely.

- Switch schema provider to postgresql
- Replace SQLite migrations with clean Postgres SQL
- Swap PrismaLibSql/PrismaBetterSqlite3 adapters for PrismaNeon
- Remove ENVIRONMENT-based branching in db.ts — single DATABASE_URL
- Delete scripts/migrate-turso.mjs and prisma/seed-turso.ts
- Update build script: prisma generate && prisma migrate deploy && next build
- Update seed.ts and docs to reflect new setup

+++ prisma/migrations/20260321000001_add_weight_to_workout_set/migration.sql
@@ -1,1 +1,1 @@
-ALTER TABLE "WorkoutSet" ADD COLUMN "weight" REAL;
+ALTER TABLE "WorkoutSet" ADD COLUMN "weight" DOUBLE PRECISION;
